i hope this works

+++ hw1port.c
@@ -3,11 +3,12 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <time.h>
+#include <stdint.h>
 
 /*generate a list of random doubles within the interval [-bound, bound], length 'size'*/
-double* generate_random_list(int size, double bound){
+double* generate_random_list(uint64_t size, double bound){
   double* ret = malloc(sizeof(double)*size);
-  for(int i=0; i<size; i++){
+  for(uint64_t i=0; i<size; i++){
     double x = (double)rand()/(double)(RAND_MAX/(2*bound)); /*generates a number in [0,2*bound]*/
       x = bound - x;
       ret[i]= x;
@@ -16,8 +17,8 @@ double* generate_random_list(int size, double bound){
 }
 
 /*update location by velocity, one time-step*/
-void update_coords(int size, double* x, double* y, double* z, double* vx, double* vy, double* vz){
-  for(int i=0; i<size; i++){
+void update_coords(uint64_t size, double* x, double* y, double* z, double* vx, double* vy, double* vz){
+  for(uint64_t i=0; i<size; i++){
     x[i] += vx[i];
     y[i] += vy[i];
     z[i] += vz[i];
@@ -45,7 +46,7 @@ int main(int argc, char** argv){
   struct timespec start,end;
   clock_gettime(CLOCK_MONOTONIC, &end);
 
-  for (int i=0; i < iters; i++) update_coords(size,x,y,z,vx,vy,vz);
+  for (uint64_t i=0; i < iters; i++) update_coords(size,x,y,z,vx,vy,vz);
   
   clock_gettime(CLOCK_MONOTONIC, &end);
